% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{inat_annotations}
\alias{inat_annotations}
\title{iNaturalist annotation term and value IDs}
\format{
A data frame with columns \code{label}, \code{term_id}, \code{term_value_id}.
}
\usage{
inat_annotations
}
\description{
A named lookup table for human-readable annotation labels and their
corresponding iNaturalist \code{term_id} and \code{term_value_id} pairs, for use
with \code{\link[=inat_fetch]{inat_fetch()}}.
}
